refactor(attachments): Migrate attachments table and remove redundant code
